move 'filtering' code in a dedicated file to separate responsibilities
